# Optimize `CoursesController#index`