--- a/app/models/course.rb
+++ b/app/models/course.rb
@@ -8,21 +8,7 @@ class Course < ApplicationRecord
 
   friendly_id :title, use: [:slugged, :finders]
 
-  def percent_completed_by(user)
-    100 * (1.0 - uncompleted_lessons_count(user) / lessons_count.to_f)
-  end
-
-  def completed_by?(user)
-    percent_completed_by(user) == 100
-  end
-
-  def lessons_count
-    lessons.count
-  end
-
-  private
-
-  def uncompleted_lessons_count(user)
-    (lessons - user.completed_lessons).count
+  def progress_for(user)
+    user.progress_for(self)
   end
 end