--- a/app/services/course_progress.rb
+++ b/app/services/course_progress.rb
@@ -1,51 +1,38 @@
 class CourseProgress
-  attr_reader :course, :user
-  private :course, :user
-
   def initialize(course, user)
     @course = course
     @user = user
   end
 
-  def self.course_started?(course, user)
-    new(course, user).course_started?
+  def started?
+    completed_lessons > 0
   end
 
-  def self.percentage_completed_by_user(course, user)
-    new(course, user).percentage_completed_by_user
+  def completed?
+    pending_lessons == 0
   end
 
-  def self.course_completed?(course, user)
-    new(course, user).course_completed?
-  end
-
-  def course_started?
-    percentage_completed_by_user > 0
-  end
-
-  def percentage_completed_by_user
-    ((100 / number_of_lessons_in_course.to_f) * number_of_lessons_completed).to_i
-  end
-
-  def course_completed?
-    percentage_completed_by_user == 100
+  def percentage
+    (100.0 * completed_lessons / total_lessons).to_i
   end
 
   private
 
-  def number_of_lessons_in_course
-    course_lessons.size
+  attr_reader :course, :user
+
+  def lesson_ids
+    @lesson_ids ||= course.lesson_ids
   end
 
-  def course_lessons
-    course.lessons
+  def total_lessons
+    @total_lessons ||= lesson_ids.size
   end
 
   def completed_lessons
-    user.completed_lessons
+    @completed_lessons ||= user.lesson_completions.where(lesson_id: lesson_ids).size
   end
 
-  def number_of_lessons_completed
-     (course_lessons & completed_lessons).size
-   end
+  def pending_lessons
+    total_lessons - completed_lessons
+  end
 end