--- a/spec/services/course_progress_spec.rb
+++ b/spec/services/course_progress_spec.rb
@@ -1,49 +1,58 @@
 require 'rails_helper'
 
+class LessonCompletions
+  def initialize(completed_lessons)
+    @completed_lessons = completed_lessons
+  end
+
+  def where(lesson_id:)
+    @completed_lessons.find_all { |lesson| lesson_id.include?(lesson) }
+  end
+end
+
 RSpec.describe CourseProgress do
   subject { CourseProgress.new(course, user) }
 
-  let(:course) { double('Course', lessons: lessons) }
-  let(:lessons) { [lesson, other_lesson_in_course] }
-  let(:lesson) { double('Lesson') }
-  let(:other_lesson_in_course) { double('Lesson') }
-  let(:user) { double('User', completed_lessons: completed_lessons) }
-  let(:completed_lessons) { [lesson] }
+  let(:lessons) { [1,2,3,4,5,6,7,8,9,10] }
+  let(:course) { double('Course', lesson_ids: lessons) }
+  let(:user) { double('User', lesson_completions: LessonCompletions.new(completed_lessons)) }
 
+  context 'when user has not started the course' do
+    let(:completed_lessons) { [] }
 
-  describe '#course_started?' do
-    it 'returns true' do
-      expect(subject.course_started?).to eql(true)
-    end
+    it { is_expected.not_to be_started }
+    it { is_expected.not_to be_completed }
 
-    context 'when user has not started the course' do
-      let(:lesson_from_other_course) { double('Lesson') }
-      let(:completed_lessons) { [lesson_from_other_course] }
-
-      it 'returns false' do
-        expect(subject.course_started?).to eql(false)
+    describe '#percentage' do
+      it 'should be at 0%' do
+        expect(subject.percentage).to eq(0)
       end
     end
   end
 
-  describe '#percentage_completed_by_user' do
-    it 'returns 50' do
-      expect(subject.percentage_completed_by_user).to eql(50)
-    end
-  end
+  context 'when user has started the course' do
+    let(:completed_lessons) { [1,2,3] }
 
-  describe '#course_completed?' do
-    it 'returns false' do
-      expect(subject.course_completed?).to eql(false)
-    end
+    it { is_expected.to be_started }
+    it { is_expected.not_to be_completed }
 
-    context 'when user has completed all lessons in the course' do
-      let(:completed_lessons) { [lesson, other_lesson_in_course] }
-
-      it 'returns true' do
-        expect(subject.course_completed?).to eql(true)
+    describe '#percentage' do
+      it 'should be at 30%' do
+        expect(subject.percentage).to eq(30)
       end
     end
   end
 
+  context 'when user has completed the course' do
+    let(:completed_lessons) { lessons }
+
+    it { is_expected.to be_started }
+    it { is_expected.to be_completed }
+
+    describe '#percentage' do
+      it 'should be at 100%' do
+        expect(subject.percentage).to eq(100)
+      end
+    end
+  end
 end