@KevinMulhern 2018-04-14
Thanks for this brilliant PR @chancancode and thank you for all the detail you provided about the decisions you made. It seems obvious now that eager loading was the problem on these end points after reading your post, but it wasn't something we even considered being the root of the problem 🙈 

The the course progress logic was a part of the code base I always thought was too complex. I'm really impressed with how you've simplified it.

Thanks for your follow up suggestions also, I'll make issues for those straight away so they can be addressed. 

In regards to denormalizing the lessons table. We have a project we want to start soon that will require significant changes to our database design. Essentially we want to introduce tracks like Frontend, Backend and Full Stack. 

That will require the ability to make custom courses and to change the relationship between sections and courses. A section will need to belong to many courses in the new design.

Database design isn't a strong part of the current maintainer teams collective skillet at the moment. Our best idea at the moment is to create join tables which will probably be very expensive performance wise.

I was wondering if you, @gitKrystan or @zvkemp would be interested in collaborating with us on the new design? we'd love to have input from you guys on it 😄

Thanks again for all the work you've done on this, 👍 LGTM