@chancancode 2018-04-14
@KevinMulhern thank you for merging the PR! I am glad you find it useful 😄 

> It seems obvious now that eager loading was the problem on these end points after reading your post

It was not obvious to me at all before digging into it, we actually tried a few things before realizing it 😄 

I just looked at the deploy on Skylight (I assume you deployed around half an hour ago – by the way, do you need help setting up [deploy tracking](https://www.skylight.io/support/advanced-setup#deploy-tracking) on Skylight?). It is probably too early to tell for sure, but it looks like typical is about a wash (or potentially a tiny bit worse), but problem responses likely have improved.

We should probably wait a bit longer before drawing conclusions (since, as we know, the logged in vs guest cases have quite different performance characteristics so a different mix of logged in/guest requests could affect the result). But, honestly I am a little surprised! (I was expecting it to do a little better than that.) It goes to show why you should always check your work on production after deploying, since a lot of variables could be different :P

It looks like, we basically traded one expensive query (fetching and loading lessons – which is easy from the query planning/execution perspective, but expensive in terms of size) for another:

<img width="1293" alt="skylight-6" src="https://user-images.githubusercontent.com/55829/38771422-0a464b80-3fd8-11e8-8950-d92fd6580b12.png">

A few things to note:

* This is actually the time of 7 queries combined (the last of the three queries in my comment above – from `course.lessons.size`), once per each course
* Also worth pointing out, the time (and allocation) on Skylight is not the raw database time, it times the end-to-end work, including time spent in Ruby (building the query inside Active Record/Arel, etc)
* I'm not exactly sure why this would be so much slower than the first query (the one returning all the lesson IDs for a given course), as they are essentially doing the same thing. In fact, the first query should be doing slightly more work. As @zvkemp pointed out, perhaps the database is configured differently, or perhaps the difference is on the Ruby side.
* This was in fact the query that is unnecessary, at least for the case where we already loaded the IDs to show the progress (i.e. for logged in users). I was thinking we would just wait for Rails to fix it (https://github.com/rails/rails/issues/32569) since it didn't seem like a big deal when I looked at it locally. However, the Rails fix probably won't land until Rails 6, and maybe we have more incentive to work around it now until Rails can fix it for us.
* As Skylight pointed out, it is a "repeated query" of sort, as we are doing 7 of them (once per course). I think there _is_ technically a way to get a similar result in a single query (do a `COUNT` with a `GROUP BY`), but it would probably require some clever refactor to find a good spot to do it.