@chancancode 2018-04-14
> I was wondering if you, @gitKrystan or @zvkemp would be interested in collaborating with us on the new design? we'd love to have input from you guys on it 😄

Of course! Happy to help where we can. Should probably be a new thread at this point :P