@KevinMulhern 2018-04-14
Thanks @chancancode, yeah I deployed around that time. I wasn't aware of deploy tracking on Skylight. I've just enabled it 😄

It seems to have comparable performance to the previous query, which isn't a problem since we cleaned up and simplified a crufty part of the code base significantly with no performance losses 🎉 

We could look into caching some of the lesson and course queries as they don't change all that often or moving some of these things into query objects.

Thanks for being willing to help us on these db issues and all the great work you've done here 😄