@chancancode 2018-04-27
Upstream Rails issue is fixed (for Rails 6)!